backend/requirements.txt
Django==3.2.9
django-cors-headers==3.11.0
djangorestframework==3.12.4
pandas==1.3.5
psycopg2==2.8.6
gunicorn==20.1.0
gevent==21.8.0
